Add truncate() to write file functions
The write file functions (writefd, and writefile) would not truncate the
files they were overwriting. If the original contents being overwritten
was larger than the contents that are overwriting the data you would have
a file that has all of the data with space after the new data.

In other words if the orignal data was of size X, the new data is of size
Y, and X > Y, then the file would consist of Y||X[(X-Y):]